requirements.txt
pandas
numpy
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
snowflake-connector-python
pyyaml